Wave 11: Fix critical autograd gradient correctness, NN module bugs, and codebase-wide issues
AUDIT-DRIVEN FIX CAMPAIGN — 29 files modified across the entire codebase.

CRITICAL: Autograd gradient correctness (17 activation/structural backward bugs):
- Log, Abs, Tanh, Sigmoid, ReLU, SiLU, Mish, GELU, Elu: replaced grad*output
  with mathematically correct derivatives
- Softmax/LogSoftmax: ReduceSum now reduces along softmax dim, not all dims
- Concat backward: split gradient along concat axis instead of full passthrough
- Slice backward: zero-pad gradient to original input shape
- Transpose backward: compute inverse permutation
- Max/Min backward: fix tensor-tensor comparison opcodes
- Removed duplicate collect_leaf_tensors call
- Fixed accumulate_grad TensorType from scalar to proper shape
- Fixed scalar_from_tensor to respect storage_offset
- BACKWARD_GRAPH_CACHE: added bounded eviction
- Fixed all backward ops creating scalar-type TensorType nodes

CRITICAL: NN module fixes:
- Conv1d/Conv3d: added missing autograd attachment (were unusable for training)
- Conv2d: added bias.requires_grad() check
- BatchNorm1d: running stats no longer carry autograd history
- PReLU: zero_grad now clears weight gradient
- TransformerBlock: safe bias unwrap with fallback
- Added rank validation for all conv/norm layers
- Removed ResidualBlock ReLU-after-residual (non-standard)

Fixes in mathemetical/logical correctness:
- ArgMax/TopK: normalize negative axis before usize cast (was usize::MAX)
- scatter_nd: proper implementation (was no-op)
- range_f32: handle negative step correctly
- ConvTranspose2d: output dim saturating_sub prevents underflow
- ReduceMean/LayerNorm/RMSNorm: division-by-zero guards
- Softplus: numerical overflow guard (x > 88)
- Shape constant folding: F32 output matches builder expectation
- Dilation param added to conv shape inference
- eliminate_noops: now actually called (was dead code)
- expand_f32: panic on invalid broadcast instead of silent wrong

Efficiency fixes:
- numel/compute_strides: i128 product prevents overflow
- View: fresh autograd_meta prevents source tensor corruption
- to_numpy: saturating arithmetic for index updates
- clip_grad_norm_: max(1e-6) instead of +1e-6 bias
- required_nodes: Vec -> HashSet for O(1) lookups
- Fusion passes: Vec<NodeId> -> HashSet<NodeId> (O(n^2) -> O(n))
- Topo sort: LIFO stack -> FIFO deque (better cache locality)
- In-place ops: increment version counter
- im2col+GEMM conv2d optimization added
- Memory planning: transitive consumer liveness analysis
- Executor: fix slice-tightening input mapping bug

+++ src/python/nn.rs
@@ -673,7 +673,7 @@ impl_nn_module!(Upsample {
     #[new]
     fn new(scale_factor: f64, mode: String) -> Self {
         Upsample {
-            inner: core_nn::upsample::Upsample::new(scale_factor, mode),
+            inner: core_nn::upsample::Upsample::new(scale_factor as f32, mode),
         }
     }
 });
